Update comment about duplicate records
I'm giving up on de-duplicating records for now.  Having a few duplicate
data points doesn't throw off the dashboards terribly and really having
this data be accurate is difficult since we unpack it from a graph and
that is, by its nature, inaccurate.

I also found that while sometimes a duplicate record is added when the
script runs at 5-minute intervals, often the record is an
almost-duplicate where a few fields are different or missing.  I took a
few passes at this and had trouble fixing it.

Reviewers: None (just changing a comment)