fix problem with bulkdata_update.q handling non-ascii data
Summary:
The topic_mapping step of bulkdata_update.q has been failing since around
Christmas.  This morning I was able to reproduce it, and the problems seemed to
be with the non-ascii data in l'hopital's rule.

Test Plan:
I uploaded the new reducer and then ran the query that has been failing.
Since this is my first dealing with unicode issues, I also tried re-parsing (in Python)
a line of UTF-8 encoded outoput of the reducer, and making sure that an equality test of
the topic_title field and the corresponding data in the ancestor_title_json succeeds.
Note what I did not test:  query the resulting data through *Hive* that would test
any aspect of the encoding.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D1018