Support box plots of performance data
Summary:
With the 1st and 3rd quartile, we can get a better idea of the
shape of performance data. I plan to use the current 5th and 95th
percentiles as the whisker endpoints (rather than using the true
outliers) for box-and-whisker plots of performance data.

Test Plan:
So far, fingers crossed. I plan to run through the report
generation process for a dummy date.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2077