Extract assessment exercises from history
Summary:
This checkin adds assessment_defs_extractor.py, a list of which
exercises are part of which assessment.  I pulled these out of the
UserAssessment records from 2013-03-11 as a basis.

The idea is that a future commit will use this whitelist in the model
building process to limit model scope to only the relevant exercises.

Test Plan:
I ran the script as described in its docstring and verified that
assessment_defs.json looked reasonable.

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D2156