add the abilitiy to output features to mirt_train_EM.py
Summary:
This change allows the MIRT training to emit a CSV dataset
and the end of training that contains a row for each question
answered containing the correctness, the MIRT model predicted
likelihood of correctness, and the abilities vector.

It's getting to be a swiss army knife of options, I know. This
was done quick-n-easy because I hope soon to to have
standardized, abstract interfaces for our data inputs, training,
and model validation.

Test Plan:
in analytics:/ebs/kadata/accmodel/mirt, I ran
  ./mirt_train_EM.py -a 1 -n 0 -f prod_responses -o mirt_emit -e -r mirt_new_epoch=74.npz

Reviewers: eliana

Reviewed By: eliana

CC: jascha

Differential Revision: http://phabricator.khanacademy.org/D2358