Make the lower-bound on the graphs zero
Summary:
The daily exercises and goals dashboard charts are never going to be
negative, but highcharts automatically scales so it shows negative portions of
the graph, which is unhelpful. This makes the lower bound zero, so it looks
better.

Test Plan:
Visit /exercises and /goals, see that the graphs look right but
don't have a negative bit on the y-axis

Reviewers: alpert, mattfaus, joshnetterfield

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D2375