properly handle null values in hive tables
Summary: report_importer was ignoring any row that had a null value in int or boolean type field. I had realised it only once whole of import failed for one of my failed examples. I don't see why we should not allow null values in int and boolean type fields. I am not sure whether error handling is still correct but seems like it is to me. Since most of the queries I have written relied on the fact that there might be null values in int columns there is nontrivial amount of rows (2%) that gets omitted by the importer currently.

Test Plan: report_importer does not increase error count when there is a null value in a int or boolean type column

Reviewers: benkomalo, jace

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2707