do not re raise the exception in report_importer since we already handle error cases
Summary: I did not spot it before but currently re raising exception is not what we want to happen. Before my change today the raise in except has served as an else case for the if that was on line 161. Since, now there is an explicit else case there is no need to reraise the exception. I might be wrong but since we are more explicit with which exception we are catching it is obvious that the error can only be a faulted type conversion.

Test Plan: none, report_importer does not break on type conversion errors

Reviewers: benkomalo, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D2716