add export of teacher geo data as a csv file
Summary:
The csv file is generated on the server and stored there. Only one file per week is generated and if it exists it is returned immediately to user. Since the dataset is quite large for a website and is expected to grow over time presenting results in the table is not going to work without server side paging. The client side data URI option for a download has potential to crash browser since resulting file is roughly 3MB in size. Only one file is ever present since each time we check if past week can be deleted.

add description to map

bump jquery version to current

Create csv version of map markers for people to download

Test Plan: teachers-students/download-map produces csv version of markers

Reviewers: chris, jace

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D2708