Make relative path absolute when creating csv file
Summary: Due to the fact that dashboards is run as an python executable where operating system runs it through python interpreter and the script is not passed to interpreter explicitly we have to deal with the fact that working directory will be different from directory our file resides in. Since when giving absolute paths to executables the working directory will be likely set to $HOME of executing user we need to adjust paths accordingly. Specifically we must make all paths dependent on script location since it defines root directory of the server. While flask deals properly with relative paths since it makes everything relative to server root the python os module does not do that and relies on current working directory to resolve relative paths.

Test Plan: download-map does not produce 500 - Internal Server Error

Reviewers: chris, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D2725