Add counting number of students to the map - popular request
Summary: This is a simple change to add number of students to the map for teachers dashboard. I do it a long way and join with student_on_date which has a value for a teacher for each student. Since each row there is unique we can simply group this table by teacher to get their class size. I still need to add this information to the map but this query takes a while to run. I will not land it before confirming it runs without breaking.

Test Plan: run the mentioned query

Reviewers: jace, chris

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D2739