Present all available data series on the graphs
Summary: This is the frontend facing bit of extra teacher metrics that I has been asked for. Quite simple changes. The conversion part does not look so nice with that many data series but without rethinking this bit entirely it is difficult to simplify it. Tell me if you want me to have a stab at doing it or is it fine the way it is.

Test Plan: go to teachers dashboard and verify that data is there

Reviewers: chris, jace

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D2846