Will only take off once downloads are finished
Summary: Checks periodically if there are any unzipped jsons in the entity kind directories

Test Plan: Not too experienced with shell scripting, so played around with a test script that includes the relevant lines, and the behavior seems to be as expected. Please feel free to let me know if this aproach is completely bogus. :)

Reviewers: jace, mattfaus

Differential Revision: http://phabricator.khanacademy.org/D2901