Add day after end_dt to weekly report. Necessary by student_teacher_count and exercise_proficiency_summary
Summary: Added day after end_dt which is necessary for two of my jobs. It is also nice to be able to have same variables in weekly and daily job for easier porting. Missing <day_after> in weekly configuration was a cause of last weekly job failure

Test Plan: those jobs don't fail

Reviewers: jace, chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D2995