Add three new entity types to pipeline
Summary: Really appreciated the documentation in https://sites.google.com/a/khanacademy.org/forge/technical/data_n/how-to-add-an-entity-kind-to-the-data-pipeline

Test Plan:
I have to backfill anyway, I'll run these before I push them

I also would like to take the opportunity of this text box to say
that any major change to the analytics machine pipeline should result
in the changer being on-call/ready to respond within 8 hours to an
emergency anytime in the next 24. Preferably. Just cause otherwise it
sucks to backfill.

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D3214