Include new class profile route to class profile visit tracking.
Summary: Since josh has rewrite of coach reports almost ready it will be beneficial for it not to destroy the tracking of coach reports visits. Include both urls to make it easy to backfill historical data if necessary.

Test Plan:
run given scripts. class_profile as well as coach/reports are counted as
a visit to class class profile.

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D3624