Add ability to set on how many different days user has to be active in order to be counted as active in the report. Add computation of hihgly_engaged of our four groups in the dashboard
Summary: Generalised a bit the reduce script so it can take number of different days on which user has to be active. I did a bit of pattern matching and replacing here. It might not make sense in all instances what I have done so it would be great to have a verification that those numbers will actually be meaningful. With long_term users computation it reaches parity with company growth. I have not tested this query thoroughly so I will postpone landing it until I have results, which might take a while.

Test Plan: run the query

Reviewers: jace, mattfaus

Reviewed By: mattfaus

Differential Revision: http://phabricator.khanacademy.org/D3069