fix a bug with registration counts on the /growth dashboard
Summary: As mentioned in the code comments, this bug was actually
created by the weirdness of using the "joined" property on UserData
as if it were the registration date.  It's not, and in fact the
UserData may have been created days before the "registered" flag
gets set to true, and then user_daily_activity will be missing a
registration activity for that event.

Test plan: Ran manually through an interative Hive cluster.

Auditors: mattfaus