Add UserData.key and .auth_emails to UserData_Info
Summary:
UserData_Info exists as a subset of UserData that can be used to quickly link the various means of identifying a user we have. Some tables use the "user" string which is something like http://google.id/1231918273409128734 (I think this is synonmous with user_id), others use the user_email which is the email that was *originally* used to create the account, but may no longer be associated with it, others use the key which is yet another opaque string.  Finally, in April 2013, we added the ability for users to link several email addresses to the same account, which is represented in auth_emails.

By adding the UserData.key and the UserData.auth_emails, we should be able to join tables easier.  For example, the UserEvent table links to the user who emitted the event with UserData.key, but there is no easy way to get at this information right now.

Although, I honestly don't understand why `is_coached` and `is_student` are on this table.

I would like to deploy this soon so I can use it for the most active coaches query, but make changes to the data pipeline is always scary.

Test Plan:
1. Start up a new EMR instance
2. Launch Hive without the init script, like: `/home/hadoop/bin/hive -d INPATH=s3://ka-mapreduce/entity_store`
3. Run the portions of ka_hive_init.q that creates the UserDataP, UserDataIncr, and UserData tables/views.
4. Run the new `CREATE EXTERNAL TABLE IF NOT EXISTS userdata_info_p` statement
5. Query this table like so, to make sure NULL values are returned for the new columns:

```
SELECT key, auth_emails
FROM userdata_info_p
WHERE dt = '2013-10-10'
LIMIT 50;
```

(I originally got this wrong by defining the columns to be the first 2, which were then grabbed the values for user and user_id)

6. Ran this query to make sure that my JSON statements were correct

```
SELECT
    get_json_object(UserData.json, '$.key'),
    get_json_object(UserData.json, '$.auth_emails')
FROM UserData
LIMIT 100;
```

Reviewers: jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D4689