Upgrade our hive_081 references to hive_0110
Summary:
As of 2013-08-02, EMR supports Hive 0.11, which I found out by starting up a new instance and it was automatically available.  Right now, we have a lot of paths hard-coded to the old version of Hive, so we need to upgrade those paths to the new one.

http://aws.amazon.com/about-aws/whats-new/2013/08/02/amazon-elastic-mapreduce-now-supports-hive-0.11.0-and-pig-0.11.1/

Test Plan: I plan on uploading this to s3 and running the modified ka_hive_init.q on my fresh EMR cluster

Reviewers: jace, jascha

Differential Revision: http://phabricator.khanacademy.org/D4711