Reduce the fetch interval back to previous levels
Turns out that increasing the interval does not really help the speed in which we download logs and it may cause problems if we have a time interval that has a much larger number of log entries, so I'm reverting.

I added a `TODO` to make fetch_logs.py multi-threaded, which I think will help a lot.

Test Plan:
Run fetch_logs.sh on the analytics machine

Auditors: james