Fix a silly file location bug
When run as a daemon on the analytics machine the current working
directory not the root of the webapp, and the app shouldn't need to be
run from its root directory. Now we can run anywhere. Anywhere.

Test Plan:
Instead of ./main.py -d, I ran dashboards/main.py -d and verified
failure before the patch and success afterwards.