Reduce number of data points in scraped GAE dashboard graph
I was using too many data points for mongo to sort through without an
index. Using an index may be the right solution, but I have been meaning
to reduce the data fetch regardless, and this solves the issue.

Also, it seems that scraping has been broken for months, so this is just
a band-aid to get the data displaying again.

Auditors: csilvers

Test Plan:
webapps/dashboard/main.py -d then visited
http://localhost:5000/gae_stats/dashboard_charts