Introduce gae_dashboard_scrape.py
Summary:
This change introduces a scrape engine for the GAE dashboard. Prior to
this we've been writing one-off scraping report scripts like
memcache_report.py. Now, we specify the metrics we're interested in, and
the scrape engine figures out what needs doing.

The goal is to make it easier to scrape data at the all-modules,
per-module, or per-version level. Eventually all existing reports should
be rewritten as thin layers on top of gae_dashboard_scrape.

This change merges memcache_report.py and instance_report.py into
ka_report.py, and uses the new framework.

Test Plan:
To make sure that the existing scraping isn't broken, I ran the
following and verified that it printed records that I expected:

  srcdir=. ./fetch_stats.sh -nv

To run the new scraper standalone:

  echo PASSWORD | gae_dashboard_scrape.py -e khanbackups@gmail.com -A s~khan-academy

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D8482