Add IP geolocation data preprocessing script to the analytics repo
Summary:
In order to determine whether to show a banner for LearnStorm Bay Area 2015 to
a user, we determine whether a user is in the bay area based on IP address.  In
order to do this efficiently, we employ a precomputed search tree across bay
area IP address ranges.

This script precomputes the search tree for determining if a user is in the bay
area and outputs it as an inlined python literal.

Test Plan:
- ensure that when the output file is moved to webapp as
  learn_storm/bay_area_ip_data.pickle, `tools/runtests.py learn_storm` passes

Reviewers: brian

Reviewed By: brian

Subscribers: jace

Differential Revision: https://phabricator.khanacademy.org/D15577