Remove bad problem logs when creating ROC graphs
Summary: As mentioned in D15835, there are a non-trivial number of problem logs that have bad `pre_attempt_prediction` data. We're not sure how this data was generated, so I decided to use this heuristic to exclude bad data when drawing these graphs. The heuristic is that the number of original predictions should be within 10% of the number of new predictions. The bad data is only from 7-01-2014 to 10-17-2014. The training data that this process is analyzing is over 8-2014 to 2-2015, so there could be some exercises with bad prediction data logged. We did not introduce any new components in 8-2014 so the ratio of original:new predictions should be either 0% or 100%.

Test Plan:
Re-run performance analysis on these data sets:

```
./src/graph_auto_trained_models.py --data /Users/mattfaus/dev/analytics/misc/performance_data/classic-2015_02_06-46210a7ead9a11e4b0480242ac110007
./src/graph_auto_trained_models.py --data /Users/mattfaus/dev/analytics/misc/performance_data/classic-2015_01_15-23aaed029cf111e49e0052258bb9aba0
```

See the the graphs no longer had horrible plots (will demonstrate in a screenshot in the comments).

Reviewers: tony, tom

Reviewed By: tom

Subscribers: jace, #data-science

Differential Revision: https://phabricator.khanacademy.org/D16107